requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
altair==5.0.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.6
colorama==0.4.6
contourpy==1.0.5
cryptography==41.0.2
cycler==0.11.0
decorator==5.1.1
distlib==0.3.7
et-xmlfile==1.1.0
filelock==3.12.2
fonttools==4.25.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.2
munkres==1.1.4
numexpr==2.8.4
numpy==1.25.0
openai==0.27.4
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230203
Pillow==9.4.0
pip==23.2.1
pipenv==2023.7.23
platformdirs==3.9.1
plotly==5.9.0
pooch==1.4.0
protobuf==4.23.4
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
rdflib==6.3.2
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
scikit-learn==1.2.2
scipy==1.10.1
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
SPARQLWrapper==2.0.0
streamlit==1.25.0
tenacity==8.2.2
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
types-pytz==2022.4.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
wheel==0.38.4
yarl==1.8.1
zipp==3.16.2
